more day 12
Seeing the end of the tunnel of day 12, I have nearly finished the group organizer

I've also spotted a bunch of ways to slim the code down and make it more readable. But I'll hold off on that until the code I _do_ have actually does its job.

I'm also tempted to ditch most of the get_spring_groups function and just let the unfuckifyers deal with the whole mess. I have to do some benchmarking to determine if it is worth it...

I have to say its very tempting to ditch 100+ lines of frantic edge-case chasing (that doesnt really work anyway, if it worked the ufuckifiers would be redundant)....